refactor: move the task sleep primitive function from sleep module to task primitives module
